Move evaluation/ to micro_sam/v1 and drop deprecation patches
