--- micro_sam/prompt_based_segmentation.py
@@ -1,13 +1,0 @@
-"""Deprecated alias for :mod:`micro_sam.v1.prompt_based_segmentation`."""
-import warnings
-import importlib
-
-warnings.warn(
-    "'micro_sam.prompt_based_segmentation' has moved to 'micro_sam.v1.prompt_based_segmentation'. "
-    "Update your imports; this alias will be removed in a future release.",
-    DeprecationWarning, stacklevel=2,
-)
-
-
-def __getattr__(name):
-    return getattr(importlib.import_module("micro_sam.v1.prompt_based_segmentation"), name)
