+++ micro_sam/sam_annotator/_widgets.py
@@ -32,7 +32,8 @@ from qtpy import QtWidgets
 from qtpy.QtCore import QObject, Signal
 from superqt import QCollapsible
 
-from .. import instance_segmentation, util
+from .. import util
+from ..v1 import instance_segmentation
 from ..v1.multi_dimensional_segmentation import (
     PROJECTION_MODES,
     get_napari_track_data,
