+++ setup.cfg
@@ -81,9 +81,9 @@ console_scripts =
     micro_sam.precompute_embeddings = micro_sam.precompute_state:main
     micro_sam.automatic_segmentation = micro_sam.v1.automatic_segmentation:main
     micro_sam.train = micro_sam.v1.training.training:main
-    micro_sam.evaluate = micro_sam.evaluation.evaluation:main
+    micro_sam.evaluate = micro_sam.v1.evaluation.evaluation:main
     micro_sam.info = micro_sam.util:micro_sam_info
-    micro_sam.benchmark_sam = micro_sam.evaluation.benchmark_datasets:main
+    micro_sam.benchmark_sam = micro_sam.v1.evaluation.benchmark_datasets:main
     
 
 # make sure it gets included in your package
